refactor(Entities): using lombok on entities and improving flow

+++ src/main/java/com/example/financialSystem/service/ExpenseService.java
@@ -1,10 +1,12 @@
 package com.example.financialSystem.service;
 
 import com.example.financialSystem.dto.requests.ExpensePatchRequest;
+import com.example.financialSystem.dto.requests.ExpenseRequest;
 import com.example.financialSystem.dto.responses.ExpenseResponse;
 import com.example.financialSystem.exception.ExpenseDuplicateException;
 import com.example.financialSystem.exception.ExpenseNotFoundException;
 import com.example.financialSystem.exception.NoChangeDetectedException;
+import com.example.financialSystem.mapper.ExpenseMapper;
 import com.example.financialSystem.model.Expense;
 import com.example.financialSystem.model.Login;
 import com.example.financialSystem.model.User;
@@ -14,70 +16,68 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ExpenseService extends UserLoggedService {
     @Autowired
     private ExpenseRepository expenseRepository;
 
-    public ExpenseResponse createExpense(Expense expense) {
+    @Autowired
+    private ExpenseMapper expenseMapper;
+
+    public ExpenseResponse createExpense(ExpenseRequest request) {
         User user = getLoggedUser().getUser();
-        expense.setUser(user);
+        Expense expense = expenseMapper.toEntity(request);
 
-        validateExpenseDate(expense);
+        validateExpenseDate(expense.getDateFinancial());
 
-        Optional<Expense> existingExpense = expenseRepository
+        expenseRepository
                 .findByUserAndTypeAndDateFinancialAndValueAndPaymentMethod(
                         user,
                         expense.getType(),
                         expense.getDateFinancial(),
                         expense.getValue(),
                         expense.getPaymentMethod()
-                );
-
-        if (existingExpense.isPresent()) {
-            throw new ExpenseDuplicateException("Expense already exists");
-        }
+                ).ifPresent(existingExpense -> {
+                    throw new ExpenseDuplicateException("Expense already exists");
+                });
 
         expenseRepository.save(expense);
         return new ExpenseResponse(expense);
     }
 
-    public ExpenseResponse editExpense(int id, Expense updatedExpense) {
+    public ExpenseResponse editExpense(int id, ExpenseRequest updatedExpense) {
         Expense existingExpense = expenseRepository.findById(id)
                 .orElseThrow(() -> new ExpenseNotFoundException(id));
 
-        validateOnwerShip(existingExpense);
-
-        validateExpenseDate(updatedExpense);
+        validateOwnerShip(existingExpense);
+        validateExpenseDate(existingExpense.getDateFinancial());
 
         ensureChanged(existingExpense, updatedExpense);
 
-        existingExpense.setType(updatedExpense.getType());
-        existingExpense.setValue(updatedExpense.getValue());
-        existingExpense.setDateFinancial(updatedExpense.getDateFinancial());
-        existingExpense.setBaseCurrency(updatedExpense.getBaseCurrency());
-        existingExpense.setPaymentMethod(updatedExpense.getPaymentMethod());
+        existingExpense.setType(updatedExpense.expenseType());
+        existingExpense.setValue(updatedExpense.value());
+        existingExpense.setDateFinancial(updatedExpense.dateFinancial());
+        existingExpense.setBaseCurrency(updatedExpense.baseCurrency());
+        existingExpense.setPaymentMethod(updatedExpense.paymentMethod());
         existingExpense.setFixed(updatedExpense.isFixed());
 
-        updatedExpense = expenseRepository.save(existingExpense);
-
-        return new ExpenseResponse(updatedExpense);
+        return new ExpenseResponse(expenseRepository.save(existingExpense));
     }
 
-    public List<Expense> listExpense() {
-        return expenseRepository.findByUser(getLoggedUser().getUser());
+    public List<ExpenseResponse> listExpense() {
+        return expenseMapper.toDtoList(expenseRepository.findByUser(getLoggedUser().getUser()));
     }
 
     public ExpenseResponse patchExpense(int id, ExpensePatchRequest patchRequest) {
         Expense existingExpense = expenseRepository.findById(id)
                 .orElseThrow(() -> new ExpenseNotFoundException(id));
 
-        validateOnwerShip(existingExpense);
+        validateOwnerShip(existingExpense);
 
-        validateExpenseDate(existingExpense);
+        validateExpenseDate(existingExpense.getDateFinancial());
 
         if (patchRequest.getValue() != null) {
             existingExpense.setValue(patchRequest.getValue());
@@ -111,7 +111,7 @@ public class ExpenseService extends UserLoggedService {
         Expense expense = expenseRepository.findById(id)
                 .orElseThrow(() -> new ExpenseNotFoundException(id));
 
-        validateOnwerShip(expense);
+        validateOwnerShip(expense);
 
         return new ExpenseResponse(expense);
     }
@@ -121,18 +121,18 @@ public class ExpenseService extends UserLoggedService {
         Expense expense = expenseRepository.findById(id)
                         .orElseThrow(() -> new ExpenseNotFoundException(id));
 
-        validateOnwerShip(expense);
+        validateOwnerShip(expense);
 
         expenseRepository.deleteById(id);
     }
 
-    public void ensureChanged(Expense existingExpense, Expense updatedExpense) {
+    public void ensureChanged(Expense existingExpense, ExpenseRequest updatedExpense) {
         boolean unchanged =
-                existingExpense.getType().equals(updatedExpense.getType()) &&
-                        existingExpense.getValue().compareTo(updatedExpense.getValue()) == 0 &&
-                        existingExpense.getDateFinancial().equals(updatedExpense.getDateFinancial()) &&
-                        existingExpense.getBaseCurrency().equals(updatedExpense.getBaseCurrency()) &&
-                        existingExpense.getPaymentMethod().equals(updatedExpense.getPaymentMethod()) &&
+                existingExpense.getType().equals(updatedExpense.expenseType()) &&
+                        existingExpense.getValue().compareTo(updatedExpense.value()) == 0 &&
+                        existingExpense.getDateFinancial().equals(updatedExpense.dateFinancial()) &&
+                        existingExpense.getBaseCurrency().equals(updatedExpense.baseCurrency()) &&
+                        existingExpense.getPaymentMethod().equals(updatedExpense.paymentMethod()) &&
                         existingExpense.isFixed() == updatedExpense.isFixed();
 
         if (unchanged) {
@@ -141,7 +141,7 @@ public class ExpenseService extends UserLoggedService {
 
     }
 
-    private void validateOnwerShip(Expense expense) {
+    private void validateOwnerShip(Expense expense) {
         Login loggedInUser = getLoggedUser();
 
         if (expense.getUser().getId() != loggedInUser.getId()) {
@@ -149,8 +149,8 @@ public class ExpenseService extends UserLoggedService {
         }
     }
 
-    private void validateExpenseDate(Expense expense) {
-        if (expense.getDateFinancial().isAfter(java.time.LocalDate.now())) {
+    private void validateExpenseDate(LocalDate date) {
+        if (date.isAfter(LocalDate.now())) {
             throw new IllegalArgumentException("You cannot create an expense with an invalid date");
         }
