+++ src/main/java/com/example/financialSystem/service/InvestmentService.java
@@ -1,10 +1,12 @@
 package com.example.financialSystem.service;
 
 import com.example.financialSystem.dto.requests.InvestmentPatchRequest;
+import com.example.financialSystem.dto.requests.InvestmentRequest;
 import com.example.financialSystem.dto.responses.InvestmentResponse;
 import com.example.financialSystem.exception.InvestmentDuplicateException;
 import com.example.financialSystem.exception.InvestmentNotFoundException;
 import com.example.financialSystem.exception.NoChangeDetectedException;
+import com.example.financialSystem.mapper.InvestmentMapper;
 import com.example.financialSystem.model.Investment;
 import com.example.financialSystem.model.Login;
 import com.example.financialSystem.model.User;
@@ -18,7 +20,9 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 
@@ -27,13 +31,16 @@ public class InvestmentService extends UserLoggedService {
     @Autowired
     private InvestmentRepository investmentRepository;
 
-    public InvestmentResponse createInvestment(Investment investment) {
+    @Autowired
+    private InvestmentMapper investmentMapper;
+
+    public InvestmentResponse createInvestment(InvestmentRequest request) {
         User user = getLoggedUser().getUser();
+
+        Investment investment = investmentMapper.toEntity(request);
         investment.setUser(user);
 
-        validateInvestmentFields(investment);
-
-        validateInvestmentDate(investment);
+        validateInvestmentDate(investment.getDateFinancial());
 
         investmentRepository
                 .findByUserAndTypeAndBrokerName(user, investment.getType(), investment.getBrokerName())
@@ -45,76 +52,79 @@ public class InvestmentService extends UserLoggedService {
                 });
 
         investment.setCurrentValue(calculateCurrentValue(investment));
+        investment.setDaysInvested((int) ChronoUnit.DAYS.between(
+                investment.getDateFinancial(),
+                LocalDate.now()
+        ));
 
-        Investment savedInvestment = investmentRepository.save(investment);
-        return new InvestmentResponse(savedInvestment);
+        return new InvestmentResponse(investmentRepository.save(investment));
     }
 
-    public InvestmentResponse editInvestment(int id, Investment updatedInvestment) {
-        Investment existingInvestment = investmentRepository.findById(id)
+    public InvestmentResponse updateInvestment(int id, InvestmentRequest request) {
+        Investment investment = investmentRepository.findById(id)
                 .orElseThrow(() -> new InvestmentNotFoundException(id));
 
-        validateOnwerShip(existingInvestment);
-        ensureChanged(existingInvestment, updatedInvestment);
 
-        existingInvestment.setType(updatedInvestment.getType());
-        existingInvestment.setValue(updatedInvestment.getValue());
-        existingInvestment.setBaseCurrency(updatedInvestment.getBaseCurrency());
-        existingInvestment.setDateFinancial(updatedInvestment.getDateFinancial());
-        existingInvestment.setActionQuantity(updatedInvestment.getActionQuantity());
-        existingInvestment.setBrokerName(updatedInvestment.getBrokerName());
-        existingInvestment.setCurrentValue(calculateCurrentValue(existingInvestment));
+        validateOwnerShip(investment);
+        validateInvestmentDate(investment.getDateFinancial());
 
-        Investment updatedData = investmentRepository.save(existingInvestment);
-        return new InvestmentResponse(updatedData);
+        ensureChanged(investment, request);
+
+        investment.setType(request.investmentType());
+        investment.setValue(request.value());
+        investment.setBaseCurrency(request.baseCurrency());
+        investment.setDateFinancial(request.dateFinancial());
+        investment.setActionQuantity(request.actionQuantity());
+        investment.setBrokerName(request.brokerName());
+        investment.setCurrentValue(calculateCurrentValue(investment));
+        investment.setDaysInvested((int) ChronoUnit.DAYS.between(investment.getDateFinancial(), LocalDate.now()));
+
+        return new InvestmentResponse(investmentRepository.save(investment));
     }
-
 
     public InvestmentResponse getInvestmentById(int id) {
         Investment investment = investmentRepository.findById(id)
                 .orElseThrow(() -> new InvestmentNotFoundException(id));
 
-        validateOnwerShip(investment);
+        validateOwnerShip(investment);
 
         return new InvestmentResponse(investment);
     }
 
     public InvestmentResponse simulateInvestment(int id, int days) {
+        if (days <= 0) {
+            throw new IllegalArgumentException("Days must be greater than zero");
+        }
+
         Investment investment = investmentRepository.findById(id)
                 .orElseThrow(() -> new InvestmentNotFoundException(id));
 
-        validateOnwerShip(investment);
+        validateOwnerShip(investment);
 
-        BigDecimal initialValue = investment.getValue();
-        InvestmentType type = investment.getType();
-        BenchMarkRate baseCurrency = investment.getBaseCurrency();
+        double annualRate = (investment.getType().getRate() != null)
+                ? investment.getType().getRate()
+                : investment.getBaseCurrency().getAnnualRate();
 
-        double rate = (type.getRate() != null) ? type.getRate() : baseCurrency.getAnnualRate();
+        BigDecimal futureValue =
+                calculateFutureValue(investment.getValue(), annualRate, days);
 
-        double dailyRate = Math.pow(1 + rate, 1.0 / 365) - 1;
+        InvestmentResponse resp = new InvestmentResponse(investment);
+        resp.setCurrentValue(futureValue);
+        resp.setDaysInvested(days);
 
-        BigDecimal growthFactor = BigDecimal.valueOf(Math.pow(1 + dailyRate, days));
-        BigDecimal futureValue = initialValue.multiply(growthFactor).setScale(4, java.math.RoundingMode.HALF_UP);
-
-        InvestmentResponse investmentDto = new InvestmentResponse(investment);
-        investmentDto.setCurrentValue(futureValue);
-
-        investmentDto.setDaysInvested(days);
-
-        return investmentDto;
+        return resp;
     }
 
-
-    public List<Investment> listInvestments() {
-        return investmentRepository.findByUser(getLoggedUser().getUser());
+    public List<InvestmentResponse> listInvestments() {
+        return investmentMapper.toDtoList(investmentRepository.findByUser(getLoggedUser().getUser()));
     }
 
     public InvestmentResponse patchInvestment(int id, InvestmentPatchRequest patchRequest) {
         Investment existingInvestment = investmentRepository.findById(id)
                 .orElseThrow(() -> new InvestmentNotFoundException(id));
 
-        validateOnwerShip(existingInvestment);
-        validateInvestmentDate(existingInvestment);
+        validateOwnerShip(existingInvestment);
+        validateInvestmentDate(existingInvestment.getDateFinancial());
 
         if (patchRequest.getInvestmentType() != null) {
             existingInvestment.setType(patchRequest.getInvestmentType());
@@ -141,6 +151,7 @@ public class InvestmentService extends UserLoggedService {
         }
 
         existingInvestment.setCurrentValue(calculateCurrentValue(existingInvestment));
+        existingInvestment.setDaysInvested((int) ChronoUnit.DAYS.between(existingInvestment.getDateFinancial(), LocalDate.now()));
 
         investmentRepository.save(existingInvestment);
 
@@ -152,37 +163,35 @@ public class InvestmentService extends UserLoggedService {
         Investment investment = investmentRepository.findById(id)
                 .orElseThrow(() -> new InvestmentNotFoundException(id));
 
-        validateOnwerShip(investment);
+        validateOwnerShip(investment);
 
         investmentRepository.deleteById(id);
     }
 
-    private BigDecimal calculateCurrentValue(Investment investment) {
-        BigDecimal initialValue = investment.getValue();
-        InvestmentType type = investment.getType();
-        BenchMarkRate baseCurrency = investment.getBaseCurrency();
-
-        double rate;
-
-        if (type.getRate() != null) {
-            rate = type.getRate();
-        } else {
-            rate = baseCurrency.getAnnualRate();
-        }
-
-        double dailyRate = Math.pow(1 + rate, 1.0 / 365) - 1;
-
-        long days = java.time.temporal.ChronoUnit.DAYS.between(
-                investment.getDateFinancial(),
-                java.time.LocalDate.now()
-        );
+    private BigDecimal calculateFutureValue(BigDecimal initialValue, double annualRate, long days) {
 
         if (days < 0) days = 0;
 
-        BigDecimal growthFactor = BigDecimal.valueOf(Math.pow(1 + dailyRate, days));
-        BigDecimal futureValue = initialValue.multiply(growthFactor);
+        double dailyRate = Math.pow(1 + annualRate, 1.0 / 365) - 1;
 
-        return futureValue.setScale(4, java.math.BigDecimal.ROUND_HALF_UP);
+        BigDecimal growthFactor = BigDecimal.valueOf(Math.pow(1 + dailyRate, days));
+
+        return initialValue.multiply(growthFactor)
+                .setScale(4, RoundingMode.HALF_UP);
+    }
+
+    private BigDecimal calculateCurrentValue(Investment investment) {
+
+        long days = ChronoUnit.DAYS.between(
+                investment.getDateFinancial(),
+                LocalDate.now()
+        );
+
+        double annualRate = (investment.getType().getRate() != null)
+                ? investment.getType().getRate()
+                : investment.getBaseCurrency().getAnnualRate();
+
+        return calculateFutureValue(investment.getValue(), annualRate, days);
     }
 
     @Scheduled(cron = "0 0 0 * * ?")
@@ -190,28 +199,17 @@ public class InvestmentService extends UserLoggedService {
         List<Investment> investments = investmentRepository.findAll();
 
         for (Investment inv : investments) {
-            inv.setDaysInvested(inv.getDaysInvested() + 1);
+            inv.setDaysInvested((int) ChronoUnit.DAYS.between(inv.getDateFinancial(), LocalDate.now()));
             inv.setCurrentValue(calculateCurrentValue(inv));
         }
         investmentRepository.saveAll(investments);
     }
 
-    private void validateInvestmentFields(Investment investment) {
-        if (investment.getBrokerName() == null || investment.getBrokerName().isBlank()) {
-            throw new IllegalArgumentException("broker name is required");
-        }
-
-        if (investment.getActionQuantity() <= 0) {
-            throw new IllegalArgumentException("Action quantity must be greater than zero");
-        }
+    private void validateInvestmentDate(LocalDate date)  {
+        if (date.isAfter(LocalDate.now())) throw new IllegalArgumentException("Investment date cannot be in the future");
     }
 
-    private void validateInvestmentDate(Investment investment) {
-        if (investment.getDateFinancial().isAfter(LocalDate.now()))
-            throw new IllegalArgumentException("Investment date cannot be in the future");
-    }
-
-    private void validateOnwerShip(Investment investment) {
+    private void validateOwnerShip(Investment investment) {
         Login loggedInUser = getLoggedUser();
 
         if (investment.getUser().getId() != loggedInUser.getUser().getId()) {
@@ -219,14 +217,14 @@ public class InvestmentService extends UserLoggedService {
         }
     }
 
-    public void ensureChanged(Investment existingInvestment, Investment updatedInvestment) {
+    public void ensureChanged(Investment existingInvestment, InvestmentRequest investmentRequest) {
         boolean unchanged =
-                existingInvestment.getType().equals(updatedInvestment.getType()) &&
-                        existingInvestment.getValue().compareTo(updatedInvestment.getValue()) == 0 &&
-                        existingInvestment.getBaseCurrency().equals(updatedInvestment.getBaseCurrency()) &&
-                        existingInvestment.getDateFinancial().equals(updatedInvestment.getDateFinancial()) &&
-                        existingInvestment.getActionQuantity() == updatedInvestment.getActionQuantity() &&
-                        existingInvestment.getBrokerName().equals(updatedInvestment.getBrokerName());
+                existingInvestment.getType().equals(investmentRequest.investmentType()) &&
+                        existingInvestment.getValue().compareTo(investmentRequest.value()) == 0 &&
+                        existingInvestment.getBaseCurrency().equals(investmentRequest.baseCurrency()) &&
+                        existingInvestment.getDateFinancial().equals(investmentRequest.dateFinancial()) &&
+                        existingInvestment.getActionQuantity() == investmentRequest.actionQuantity() &&
+                        existingInvestment.getBrokerName().equals(investmentRequest.brokerName());
 
         if (unchanged) {
             throw new NoChangeDetectedException("No changes in this investment");
